backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[argon2]
argon2-cffi
cryptography
redis
python-multipart
email-validator
pydantic-extra-types
deepagents
chromadb
paddleocr
langchain
langchain-mistralai
langchain-anthropic